fix: cancel-triggered shifts, spread-preserving price movement, and rebalanced HLR curves
Cancel events that deplete the best level now trigger a price shift,
matching the HLR2014 model where any queue depletion causes a price
change. Previously only executions could shift, leaving cancelled-to-zero
levels as zombies that silently dropped subsequent executes.

Shifts now move both sides' prices together (bid shift also moves ask
down, ask shift also moves bid up) to maintain constant spread, instead
of diverging bid/ask indefinitely.

Added guard for num_levels_==1 to prevent size_t underflow in shift
functions.

Rebalanced HLR2014 default curves: reduced add rate (5.0 -> 2.5), made
cancel rate linear in queue size (0.15n), and made market order rate
roughly constant (~1.0) instead of decaying to zero. Equilibrium queue
depth is now ~10 instead of ~220, producing regular price shifts.

Fixed MSVC printf format warnings in test_producer.cpp (%lu -> %llu).

+++ tests/qrsdp/test_producer.cpp
@@ -191,8 +191,8 @@ TEST(QrsdpProducer, TraceShiftAndPrintFirst20) {
         const int32_t prev_bid = replayBook.bestBid().price_ticks;
         const int32_t prev_ask = replayBook.bestAsk().price_ticks;
         if (i < print_limit) {
-            std::printf("[%zu] ts_ns=%lu type=%u side=%u price=%d qty=%u order_id=%lu\n",
-                        i, rec.ts_ns, rec.type, rec.side, rec.price_ticks, rec.qty, rec.order_id);
+            std::printf("[%zu] ts_ns=%llu type=%u side=%u price=%d qty=%u order_id=%llu\n",
+                        i, (unsigned long long)rec.ts_ns, rec.type, rec.side, rec.price_ticks, rec.qty, (unsigned long long)rec.order_id);
         }
         SimEvent ev = recordToSimEvent(rec);
         replayBook.apply(ev);
@@ -207,8 +207,8 @@ TEST(QrsdpProducer, TraceShiftAndPrintFirst20) {
         EXPECT_GE(ask.price_ticks - bid.price_ticks, 1) << "spread >= 1 after event " << i;
     }
 
-    std::printf("events_written=%lu close_ticks=%d shift_count=%d\n",
-                result.events_written, result.close_ticks, shift_count);
+    std::printf("events_written=%llu close_ticks=%d shift_count=%d\n",
+                (unsigned long long)result.events_written, result.close_ticks, shift_count);
 }
 
 // --- HLR2014 Model I (curve intensity) tests ---
